+++ react/tic-tac-toe/src/app/screens/Game/components/Board/index.js
@@ -5,65 +5,20 @@ import Square from '../Square';
 import styles from './styles.module.scss';
 
 class Board extends Component {
-  state = {
-    squares : Array(9).fill(null),
-    xIsNext: true
-  };
-
-  handleClick = (i) => {
-    const squares = this.state.squares.slice();
-    if (this.calculateWinner() || squares[i]) {
-      return;
-    }
-    squares[i] = this.state.xIsNext ? 'X' : 'O';
-    this.setState({
-      squares : squares,
-      xIsNext: !this.state.xIsNext
-    });
-  };
 
   renderSquare(i) {
     return (
       <Square
-        value={this.state.squares[i]}
-        onClick = {() => this.handleClick(i)}
+        value={this.props.squares[i]}
+        onClick = {() => this.props.onClick(i)}
       />
     );
   }
 
-  calculateWinner = () => {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (this.state.squares[a] && 
-      this.state.squares[a] === this.state.squares[b] && 
-      this.state.squares[a] === this.state.squares[c]) {
-        return this.state.squares[a];
-      }
-    }
-    return null;
-  }
 
   render() {
-    const winner = this.calculateWinner();
-    let status;
-    if(winner){
-      status = 'Winner: ' + winner;
-    } else{
-      status = 'Next player: ' + (this.state.xIsNext ? 'X' : 'O');
-    }
     return (
       <div>
-        <div className={styles.status}>{status}</div>
         <div className={styles.boardRow}>
           {this.renderSquare(0)}
           {this.renderSquare(1)}
